Improve Flow type (#160)
* Update dependencies to @latest

* Update README.md with correct types for exports

* Improve Flow coverage across the board

* Rename src/flow.js to src/types.js

* Fixes to non-transpiled files (Comment syntax for Flow) and minor fix to type in browser.js

+++ fusion-plugin-i18n/chunk-translation-map.js
@@ -3,38 +3,61 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @noflow
+ * @flow
  */
+
 /* eslint-env node */
 
 class ChunkTranslationMap {
+  /*:: translations: Map<number | string, Map<string, Set<string>>>; */
+
   constructor() {
     this.translations = new Map();
   }
 
-  add(filename, chunkIds, translations) {
+  add(
+    filename /*: string */,
+    chunkIds /*: Array<number | string> */,
+    translations /*: Array<string> */
+  ) /*: void */ {
     for (const chunkId of chunkIds) {
       if (!this.translations.has(chunkId)) {
         this.translations.set(chunkId, new Map());
       }
 
       const translationsForChunk = this.translations.get(chunkId);
+      if (!translationsForChunk)
+        throw new Error('translations missing chunkId');
+
       for (const key of translations) {
         if (!translationsForChunk.has(key)) {
           translationsForChunk.set(key, new Set());
         }
 
         const filesForTranslation = translationsForChunk.get(key);
+        if (!filesForTranslation)
+          throw new Error('unable to get Set() in translations for chunk');
+
         filesForTranslation.add(filename);
       }
     }
   }
 
-  dispose(filename, chunkIds, translations) {
+  dispose(
+    filename /*: string */,
+    chunkIds /*: Array<number> */,
+    translations /*: Array<string> */
+  ) /*: void */ {
     for (const chunkId of chunkIds) {
       const translationsForChunk = this.translations.get(chunkId);
+      if (!translationsForChunk)
+        throw new Error('translations missing chunkId');
+
       for (const key of translations) {
         const filesForTranslation = translationsForChunk.get(key);
+        if (!filesForTranslation)
+          throw new Error('unable to get Set() in translations for chunk');
+
         filesForTranslation.delete(filename);
 
         if (filesForTranslation.size === 0) {
@@ -44,11 +67,14 @@ class ChunkTranslationMap {
     }
   }
 
-  translationsForChunk(chunkId) {
+  translationsForChunk(chunkId /*: number | string */) /*: Iterator<string> */ {
     if (!this.translations.has(chunkId)) {
-      return new Set();
+      return new Set().keys();
     }
-    return this.translations.get(chunkId).keys();
+    const translationsForChunk = this.translations.get(chunkId);
+    if (!translationsForChunk) throw new Error('translations missing chunkId');
+
+    return translationsForChunk.keys();
   }
 }
